Refactor Campaign component
Moved Campaign component to own subfolder, created campaign constants
