+++ frontend/src/components/Campaign/Campaign.config.ts
@@ -1,0 +1,3 @@
+export const CAMPAIGN_TITLE_PLACEHOLDER = "Campaign Title";
+export const CAMPAIGN_DESCRIPTION_PLACEHOLDER = "Enter a position description including key selection criteria, experience, etc";
+
